move create task logic into parent class
