+++ MLAgentsProject/Assets/MLAgentsProject/Scripts/Database/Tasks/OllamaProofTask.cs
@@ -4,14 +4,9 @@ using System.Diagnostics;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using UnityEngine;
 
 public class OllamaProofTask : BaseTask<OllamaProofTask, Response>
 {
-    public OllamaProofTask() => Application.quitting += OnApplicationQuit;
-
-    private void OnApplicationQuit() => CancellationTokenSource.Cancel();
-
     public override async Task Process(MessageList messageList, string jsonDataFilename)
     {
         InitializeCounters();
@@ -65,6 +60,49 @@ public class OllamaProofTask : BaseTask<OllamaProofTask, Response>
         {
             stopwatch.Stop();
             Log.Message($"Generate: Proof task for user content '{userContent}' completed in {stopwatch.ElapsedMilliseconds} ms.");
+        }
+    }
+
+    public override List<Task> CreateTasks(MessageList messageList, List<Message> newMessagesList, List<Message> errorMessageList, int totalMessages)
+    {
+        Log.Message("OllamaProofTask: Creating tasks for proof processing.");
+        return Enumerable.Range(0, totalMessages)
+                         .Select(i =>
+                         {
+                             var message = GetMessage(messageList, i);
+                             var userContent = GetUserContent(message);
+                             var agentContent = GetAgentContent(message);
+                             return ProcessContent(userContent, agentContent, message, newMessagesList, errorMessageList, i, totalMessages);
+                         })
+                         .ToList();
+    }
+
+    public override async Task ProcessContent(string userContent, string agentContent, Message message, List<Message> newMessagesList, List<Message> errorMessageList, int index, int totalMessages)
+    {
+        await Semaphore.WaitAsync(CancellationTokenSource.Token);
+        try
+        {
+            Log.Message($"OllamaProofTask: Processing user content: {userContent} ({index + 1} of {totalMessages})");
+
+            var responses = await Execute(userContent, agentContent, TimeSpan.FromSeconds(30), 10);
+            ValidateResponses(responses.Select(r => r.ToString()).ToList(), userContent);
+
+            // New log line specific to proof task
+            Log.Message("OllamaProofTask: Validation completed for proof task.");
+
+            AddNewMessages(message, responses.Select(r => r.ToString()).ToList(), newMessagesList);
+            UpdateCounters(responses.Count, newMessagesList.Count);
+
+            Log.Message($"OllamaProofTask: Generated {responses.Count} responses for user content: {userContent}. Completed {index + 1} of {totalMessages} tasks.");
+        }
+        catch (Exception ex)
+        {
+            Log.Message($"OllamaProofTask: Error processing user content '{userContent}': {ex.Message}");
+            AddErrorMessage(message, errorMessageList);
+        }
+        finally
+        {
+            Semaphore.Release();
         }
     }
 }
